Convert CustomServerDialog to TypeScript
Refactored; TypeScript; functional component.
