+++ src/components/views/auth/CustomServerDialog.tsx
@@ -1,0 +1,45 @@
+/*
+Copyright 2015, 2016 OpenMarket Ltd
+Copyright 2019, 2020 The Matrix.org Foundation C.I.C.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import React from 'react';
+import { _t } from '../../../languageHandler';
+import SdkConfig from '../../../SdkConfig';
+
+export function CustomServerDialog () {
+    const brand = SdkConfig.get().brand;
+    return (
+        <div className="mx_ErrorDialog">
+            <div className="mx_Dialog_title">
+                { _t("Custom Server Options") }
+            </div>
+            <div className="mx_Dialog_content">
+                <p>{_t(
+                    "You can use the custom server options to sign into other " +
+                    "Matrix servers by specifying a different homeserver URL. This " +
+                    "allows you to use %(brand)s with an existing Matrix account on a " +
+                    "different homeserver.",
+                    { brand },
+                )}</p>
+            </div>
+            <div className="mx_Dialog_buttons">
+                <button onClick={this.props.onFinished} autoFocus={true}>
+                    { _t("Dismiss") }
+                </button>
+            </div>
+        </div>
+    );
+}
